pyproject.toml
[project]
name = "mememind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.7.0",
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "boto3>=1.38.19",
    "celery>=5.5.2",
    "chromadb>=1.0.11",
    "fastapi[standard]>=0.115.9",
    "gradio>=5.33.0",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "loguru>=0.7.3",
    "lxml>=5.4.0",
    "pandas>=2.3.0",
    "pydantic-settings>=2.9.1",
    "python-docx>=1.1.2",
    "redis>=6.1.0",
    "sqlalchemy>=2.0.41",
    "torch>=2.7.0",
    "transformers>=4.52.3",
    "unstructured[docx,md,pdf,pptx,xlsx]>=0.17.2",
]
